implement stripe payments

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { CheckoutService } from '../checkout.service';
 import { FormGroup } from '@angular/forms';
 import { BasketService } from 'src/app/basket/basket.service';
@@ -6,20 +6,62 @@ import { ToastrService } from 'ngx-toastr';
 import { Basket } from 'src/app/shared/models/basket';
 import { Address } from 'src/app/shared/models/user';
 import { NavigationExtras, Router } from '@angular/router';
+import { Stripe, StripeCardCvcElement, StripeCardExpiryElement, StripeCardNumberElement, loadStripe } from '@stripe/stripe-js';
 
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent {
+export class CheckoutPaymentComponent implements OnInit {
 
   @Input() checkoutForm?: FormGroup;
+  @ViewChild('cardNumber') cardNumberElement?: ElementRef;
+  @ViewChild('cardExpiry') cardExpiryElement?: ElementRef;
+  @ViewChild('cardCvc') cardCvcElement?: ElementRef;
+
+  stripe: Stripe | null = null;
+  cardNumber?: StripeCardNumberElement;
+  cardExpiry?: StripeCardExpiryElement;
+  cardCvc?: StripeCardCvcElement;
+  cardErrors: any;
 
   constructor(private checkoutService: CheckoutService,
     private basketService: BasketService,
     private toastr: ToastrService,
     private router: Router) {}
+
+  ngOnInit(): void {
+    loadStripe(
+      'pk_test_51Mv8veKKkVkvo185clexhw9XwhoPJzmeqgWscARC2Zm8kKSl3Up1f0X4fmsrqyFJuqueJBnHDhmTp8DPEp0ZOLmu00YtULI8Pg')
+        .then(stripe => {
+          this.stripe = stripe;
+          const elements = stripe?.elements();
+
+          if (elements) {
+            this.cardNumber = elements.create('cardNumber');
+            this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+            this.cardNumber.on('change', event => {
+              if (event.error) this.cardErrors = event.error.message;
+              else this.cardErrors = null;
+            });
+
+            this.cardExpiry = elements.create('cardExpiry');
+            this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+            this.cardExpiry.on('change', event => {
+              if (event.error) this.cardErrors = event.error.message;
+              else this.cardErrors = null;
+            });
+
+            this.cardCvc = elements.create('cardCvc');
+            this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+            this.cardCvc.on('change', event => {
+              if (event.error) this.cardErrors = event.error.message;
+              else this.cardErrors = null;
+            });
+          }
+        })
+  }
 
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
@@ -32,10 +74,24 @@ export class CheckoutPaymentComponent {
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: order => {
         this.toastr.success('Order created successfully!');
-        this.basketService.deleteLocalBasket();
-        
-        const navigationExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigationExtras)
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!, // this will also grab expiry and cvc
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then( result => {
+          console.log(result);
+
+          if (result.paymentIntent) {
+            this.basketService.deleteLocalBasket();     
+            const navigationExtras: NavigationExtras = {state: order};
+            this.router.navigate(['checkout/success'], navigationExtras);
+          } else {
+            this.toastr.error(result.error.message);
+          }
+        });
       }
     });
   }
